+++ owlcms/src/main/java/app/owlcms/displays/scoreboard/Results.java
@@ -860,7 +860,7 @@ public class Results extends PolymerTemplate<TemplateModel>
 
 	protected void getAthleteJson(Athlete a, JsonObject ja, Category curCat, int liftOrderRank, FieldOfPlay fop) {
 		String category;
-		category = curCat != null ? curCat.getName() : "";
+		category = curCat != null ? curCat.getComputedName() : "";
 		ja.put("fullName", a.getFullName() != null ? a.getFullName() : "");
 		ja.put("teamName", a.getTeam() != null ? a.getTeam() : "");
 		ja.put("yearOfBirth", a.getYearOfBirth() != null ? a.getYearOfBirth().toString() : "");