+++ owlcms/src/main/java/app/owlcms/nui/results/ResultsContent.java
@@ -549,7 +549,7 @@ public class ResultsContent extends AthleteGridContent implements HasDynamicTitl
 		genderFilter.setPlaceholder(getTranslation("Gender"));
 		genderFilter.setItems(Gender.M, Gender.F);
 		genderFilter.setItemLabelGenerator((i) -> {
-			return i == Gender.M ? getTranslation("Gender.M") : getTranslation("Gender.F");
+			return i == Gender.M ? getTranslation("Gender.Men") : getTranslation("Gender.Women");
 		});
 		genderFilter.setClearButtonVisible(true);
 		genderFilter.addValueChangeListener(e -> {