+++ src/main/markdown/ReleaseNotes.md
@@ -3,7 +3,9 @@
 - 38.1 Enhancements
   - The attempt board now shows the athlete category and the lift type.  If you only want to see the attempt number, leave the `AttemptBoard_lift_attempt_number` translation empty.
   - The announcer "refresh list" button has been renamed to "Reload Group". It reloads the group completely (same as exiting and re-entering the group). This also sends a refresh signal to all displays. This is useful if for some reason it is necessary to edit athlete registration information.
+  - The gender letters used for displaying *age groups* and *categories* are no longer fixed to `M` and `F` and can now be translated (for example, Germany could choose to use U17 D instead of U17 F)
 - 38.1 Fixes
+  - 38.1.0-alpha02: A specific error message is given if the athlete's body weight is changed and the athlete does not meet the minimum entry total for the new category and is therefore not eligible.
   - 38.1.0-alpha01: the current attempt number was not being hidden correctly on the down signal.
   - The total of the last snatch athlete was being shown during the clean & jerk break on the "current athlete" video streaming footer.
 - 38.0 Enhancements